pyproject.toml
[project]
name = "api-manager-orders"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.120.3",
    "greenlet>=3.2.4",
    "jose>=1.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pyproject-toml>=0.1.0",
    "sqlalchemy>=2.0.44",
]
